third_party/blink/web_tests/accessibility/aria-hidden-with-elements.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body id="body">

<div id="container" aria-label="container">

  <table border=1 summary="test" aria-hidden="true">
    <tr><td>cell</td><td>cell</td><td>cell</td></tr>
    <tr><td>cell</td><td>cell</td><td>cell</td></tr>
  </table>

  <div role="grid" aria-hidden="true">
    <div role="row">
      <div role="gridcell">cell</div>
    </div>
  </div>

  <ul aria-hidden="true">
  <li>test
  <li>test
  </ul>

  <div aria-label="wrapper">
    <div role="listbox" aria-hidden="true">
      <div role="option">option</div>
      <div role="option">option</div>
    </div>
  </div>

  <h1>heading</h1>

</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that aria-hidden works as expected on elements that are subclasses of AccessibilityRenderObject");

   if (window.accessibilityController) {



       var container = accessibilityController.accessibleElementById("container");

       // first child is an AXGroup that has no children
       var child = container.childAtIndex(0);
       shouldBe("child.childrenCount", "0");

       // The second child we ask for from the web area should be the heading, and we should have skipped over everything else.
       child = container.childAtIndex(1);
       shouldBe("child.role", "'AXRole: AXHeading'");
    }

</script>

</body>
</html>
